move MultiEmitter in its own file
